refactor(accelerator): q7e3-F-12 move server.rs tests to server/tests.rs
server.rs read as a 1424-LOC monolith though its production code is ~330
LOC — the rest was a colocated test module. The audit established these are
router-level INTEGRATION tests (each builds router() + oneshots a request),
not submodule unit tests, so scattering them into prove/auth/host unit-test
mods would be wrong; the right home is a dedicated server-integration file.
Extract the whole #[cfg(test)] mod into server/tests.rs via 'mod tests;' —
a file module sits at the same module-tree position as the inline one, so
every 'super::*' path is unchanged (verbatim move, zero import edits).
server.rs 1424->331 LOC; clippy -D warnings clean; core 132/132 unchanged.
